Transformation: move API model to api package
