+++ transformation/src/interfaces/jobResult.ts
@@ -1,0 +1,8 @@
+import Stats from '../pipeline-execution/stats'
+import JobError from '@/pipeline-execution/sandbox/jobError'
+
+export default interface JobResult {
+  data?: object;
+  error?: JobError;
+  stats: Stats;
+}
